requirements.txt
attrs==19.3.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
descartes==1.1.0
Fiona==1.8.11
geopandas==0.6.2
kiwisolver==1.1.0
matplotlib==3.1.2
maxminddb==1.5.1
maxminddb-geolite2==2018.703
munch==2.5.0
numpy==1.17.4
pandas==0.25.3
pyparsing==2.4.5
pyproj==2.4.1
python-dateutil==2.8.1
python-geoip==1.2
python-geoip-geolite2==2015.303
pytz==2019.3
Shapely==1.6.4.post2
six==1.13.0
